requirements.txt
selenium==4.9.0
beautifulsoup4==4.12.3
webdriver_manager==4.0.1
lxml==5.2.2
python-dotenv==1.0.1
twilio==9.2.1
undetected-chromedriver==3.5.5
boto3==1.34.138
supabase==2.5.1
PyJWT==2.8.0
sendgrid==6.11.0
selenium-stealth==1.0.6
unidecode==1.3.8